# docs/swagger.yaml
basePath: /
definitions:
...
  models.Bahan:
    properties:
      batas_minimum:
        type: number
      harga_saat_ini:
        description: Update otomatis dari pembelian terakhir
        type: number
      id:
        type: integer
      nama_bahan:
        type: string
      satuan:
        description: gr, ml, pcs
        type: string
      stok:
        type: number
      urutan:
        type: integer
    type: object
...
  models.Barang:
    properties:
      Urutan:
        type: integer
      hargaDefault:
        format: float64
        type: number
      id:
        type: integer
      kebutuhan_adonan:
        description: Berapa gram / fraksi per 1 roti
        type: number
      kemasan_detail:
        items:
          $ref: '#/definitions/models.BarangKemasan'
        type: array
      masa_simpan:
        description: Default 2 hari
        type: integer
      metode_konversi:
        description: '"Gram" atau "Pcs"'
        type: string
      namaBarang:
        type: string
      resep:
        $ref: '#/definitions/models.Resep'
      resep_id:
        description: TAMBAHAN MODUL INVENTORY
        type: integer
    type: object
  models.BarangInput:
    properties:
      HargaDefault:
        example: 15000
        type: number
      NamaBarang:
        example: Roti Tawar Spesial
        type: string
      kebutuhan_adonan:
        example: 0.5
        type: number
      kemasan_detail:
        items:
          $ref: '#/definitions/models.KemasanInput'
        type: array
      masa_simpan:
        example: 3
        type: integer
      metode_konversi:
        example: BAGI
        type: string
      resep_id:
        example: 1
        type: integer
    type: object
  models.BarangKemasan:
    properties:
      bahan:
        $ref: '#/definitions/models.Bahan'
      bahan_id:
        type: integer
      barang_id:
        type: integer
      id:
        type: integer
      kebutuhan:
        description: (Bisa pecahan, misal 0.25)
        type: number
    type: object
...
  models.KemasanInput:
    properties:
      bahan_id:
        example: 2
        type: integer
      kebutuhan:
        example: 1.5
        type: number
    type: object
...
  models.ProfilTiara:
    properties:
      alamat:
        type: string
      id:
        type: integer
      logoPath:
        type: string
      nama:
        type: string
      noHP:
        type: string
      noTelp:
        type: string
    type: object
...
  models.Resep:
    properties:
      bahan_detail:
        items:
          $ref: '#/definitions/models.ResepBahan'
        type: array
      id:
        type: integer
      nama_resep:
        type: string
      target_gramasi:
        description: Total adonan matang dr 1 Resep
        type: number
    type: object
  models.ResepBahan:
    properties:
      bahan:
        $ref: '#/definitions/models.Bahan'
      bahan_id:
        type: integer
      id:
        type: integer
      kebutuhan:
        description: Butuh berapa gr/ml/pcs
        type: number
      resep_id:
        type: integer
    type: object
...
  models.Toko:
    properties:
      IsHarian:
        type: boolean
      alamat:
        type: string
      id:
        type: integer
      namaToko:
        type: string
      noTelp:
        type: string
      siklusDua:
        type: boolean
      siklusJumatSelasa:
        type: boolean
      siklusKamisSenin:
        description: Flag Siklus
        type: boolean
      siklusSabtuRabu:
        type: boolean
    type: object
...
  models.UrutanBarangInput:
    properties:
      id:
        example: 1
        type: integer
      urutan:
        example: 2
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    name: Aristo Budiman
  description: Dokumentasi API Internal untuk Sistem Inventory, Nota, dan Kas.
  title: Tiara Bakery Master API
  version: "1.0"
paths:
...
  /api/barangs:
    get:
      consumes:
      - application/json
      description: Mengambil daftar master barang lengkap dengan relasi resep dan
        kemasan, diurutkan berdasarkan urutan tampilan.
      produces:
      - application/json
      responses:
        "200":
          description: List data barang berhasil ditarik
          schema:
            items:
              $ref: '#/definitions/models.Barang'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil Seluruh Master Barang
      tags:
      - 03. Master Barang
    post:
      consumes:
      - application/json
      description: Menyimpan data master barang baru beserta kebutuhan kemasannya
        (kardus/plastik) jika ada.
      parameters:
      - description: Data Barang utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BarangInput'
      produces:
      - application/json
      responses:
        "200":
          description: Barang dan Kemasan berhasil dibuat
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON tidak valid
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Gagal menyimpan ke database
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat Master Barang Baru
      tags:
      - 03. Master Barang
  /api/barangs/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus master barang dari daftar aktif (masuk ke tong sampah).
      parameters:
      - description: ID Barang
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Barang berhasil dihapus
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus Master Barang (Soft Delete)
      tags:
      - 03. Master Barang
    put:
      consumes:
      - application/json
      description: Memperbarui data master barang beserta list kemasannya berdasarkan
        ID.
      parameters:
      - description: ID Barang
        in: path
        name: id
        required: true
        type: integer
      - description: Data Update Barang utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BarangInput'
      produces:
      - application/json
      responses:
        "200":
          description: Barang berhasil diupdate
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON tidak valid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Barang tidak ditemukan
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Master Barang
      tags:
      - 03. Master Barang
  /api/barangs/reorder:
    put:
      consumes:
      - application/json
      description: Memperbarui posisi/urutan barang untuk tampilan di form pembuatan
        nota. (Dipanggil saat Drag & Drop di UI Vue)
      parameters:
      - description: 'Array of object dengan key: id, urutan'
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/models.UrutanBarangInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Urutan barang berhasil diperbarui
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format data salah
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Urutan Tampilan Barang
      tags:
      - 03. Master Barang
...
  /api/profil:
    get:
      consumes:
      - application/json
      description: Mengambil data profil Tiara Bakery (Nama dan Alamat) untuk keperluan
        header nota PDF.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menarik data profil
          schema:
            $ref: '#/definitions/models.ProfilTiara'
      security:
      - BearerAuth: []
      summary: Ambil Profil Perusahaan
      tags:
      - 02. Master Profil
...
  /api/sampah:
    get:
      consumes:
      - application/json
      description: Mengambil semua data master (toko, barang, bahan, resep) yang berstatus
        soft-deleted.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menarik data sampah
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil Data Tong Sampah
      tags:
      - 07. Master Sampah
  /api/sampah/{type}/{id}:
    put:
      consumes:
      - application/json
      description: 'Mengembalikan data master yang ada di tong sampah ke daftar aktif.
        Parameter type bisa berisi: ''toko'', ''barang'', ''bahan'', atau ''resep''.'
      parameters:
      - description: Jenis Data (toko/barang/bahan/resep)
        in: path
        name: type
        required: true
        type: string
      - description: ID Data
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Data berhasil dipulihkan
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Jenis data tidak valid
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Pulihkan Data Terhapus
      tags:
      - 07. Master Sampah
  /api/tokos:
    get:
      consumes:
      - application/json
      description: Mengambil data seluruh mitra toko beserta pengaturan siklus tagihannya.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menarik data toko
          schema:
            items:
              $ref: '#/definitions/models.Toko'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil Seluruh Master Toko
      tags:
      - 04. Master Toko
    post:
      consumes:
      - application/json
      description: Menyimpan data mitra toko baru beserta aturan siklus tagihannya.
      parameters:
      - description: Data Toko Baru
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Toko'
      produces:
      - application/json
      responses:
        "200":
          description: Toko berhasil dibuat
          schema:
            $ref: '#/definitions/models.Toko'
        "400":
          description: Format JSON tidak valid
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Gagal menyimpan ke database
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buat Master Toko Baru
      tags:
      - 04. Master Toko
  /api/tokos/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus mitra toko dari daftar aktif (masuk ke tong sampah).
      parameters:
      - description: ID Toko
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Toko berhasil dihapus
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hapus Master Toko (Soft Delete)
      tags:
      - 04. Master Toko
    put:
      consumes:
      - application/json
      description: Memperbarui profil mitra toko dan aturan siklus tagihannya berdasarkan
        ID.
      parameters:
      - description: ID Toko
        in: path
        name: id
        required: true
        type: integer
      - description: Data Update Toko
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Toko'
      produces:
      - application/json
      responses:
        "200":
          description: Toko berhasil diupdate
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Format JSON tidak valid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Toko tidak ditemukan
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Master Toko
      tags:
      - 04. Master Toko
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: "Bearer {token_jwt}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"